Break out TimelineClient
